feat: Use the URL to navigate between pages
Instead of using showSearchPage state variable to keep track of which
page we're on, React Router was added to use the URL in the browser's
address bar. This will ensure that users can use the browser's back and
forward buttons to navigate between pages, as well as provide a good URL
they can bookmark and share.